+++ docs/source/changelog.rst
@@ -11,6 +11,7 @@ ChangeLog
 ***********
 
 * Allow the creation of Certificates with multiple OUs in their subject (command-line only).
+* Fix issues with handling CAs with a password on the command-line.
 * Django 2.1 is now fully supported.
 * Fix example command to generate a CSR (had a stray '/').